Remove old AreaGraph component
