+++ web/src/scss/content/panel/_country-panel.scss
@@ -143,7 +143,7 @@
     }
 }
     
-#country-history-mix, #country-history-mix-react {
+#country-history-mix-react {
     height: 10rem;
 
     .layer path {
